AdlibGold: More translations from DOSBox Staging

+++ src/Spice86.Core/Emulator/Devices/Sound/MathUtils.cs
@@ -1,6 +1,6 @@
 namespace Spice86.Core.Emulator.Devices.Sound;
 
 public static class MathUtils {
-    public static double decibel_to_gain(double decibel) =>
-        Math.Pow(10.0f, decibel / 20.0f);
+    public static float decibel_to_gain(float decibel) =>
+        (float)Math.Pow(10.0f, decibel / 20.0f);
 }
